Adjust Board code locations. Use Dialog component for Board item creation flow.
