+++ src/translations/index.ts
@@ -22,6 +22,7 @@ export default {
     hello: "Hello",
     "add lane": "Add lane",
     "add board item": "Add item to board",
+    "edit board item": "Edit item",
     "board lane": "Board lane",
     "board item": "Board item",
     "name lane": "Name this lane…",
@@ -32,12 +33,15 @@ export default {
     "lane options": "Lane options",
     "sort-order alphabetical descending": "A-Z",
     "sort-order alphabetical ascending": "Z-A",
+    cancel: "Cancel",
+    confirm: "Confirm",
   } as TTranslations,
   fa: {
     locale: "fa",
     hello: "سلام",
     "add lane": "خط اضافه کنید",
     "add board item": "مورد را به تخته اضافه کنید",
+    "edit board item": "ویرایش آیتم",
     "board lane": "خط تخته",
     "board item": "مورد هیئت مدیره",
     "name lane": "این خط را نامگذاری کنید…",
@@ -48,5 +52,7 @@ export default {
     "lane options": "گزینه های خط",
     "sort-order alphabetical descending": "A-Z",
     "sort-order alphabetical ascending": "Z-A",
+    cancel: "لغو",
+    confirm: "تایید",
   } as TTranslations,
 };
